+++ src/llmcompressor/transformers/__init__.py
@@ -6,6 +6,7 @@ Tools for integrating LLM Compressor with transformers training flows
 
 # isort: skip_file
 # (import order matters for circular import avoidance)
+from .wrap import wrap_hf_model_class
 from .utils import *
-from .sparsification import SparseAutoModel, SparseAutoModelForCausalLM, wrap_hf_model_class
+from .sparsification import SparseAutoModel, SparseAutoModelForCausalLM
 from .finetune import *
